+++ packages/framework/src/Commands/HydeBuildSearchCommand.php
@@ -61,12 +61,7 @@ class HydeBuildSearchCommand extends Command
         $this->comment('Generating search page...');
         file_put_contents(
             Hyde::path('_site/'.config('docs.output_directory', 'docs').'/search.html'),
-            view('hyde::layouts.docs')->with([
-                'page' => new DocumentationPage([], '', 'Search', 'search'),
-                'title' => 'Search',
-                'markdown' => view('hyde::pages.documentation-search')->render(),
-                'currentPage' => ''.config('docs.output_directory', 'docs').'/search',
-            ])->render()
+            view('hyde::pages.documentation-search')->render()
         );
 
         $this->line(' > Created <info>_site/'.config('docs.output_directory', 'docs').'/search.html</> in '.